+++ src/WayPoint/WayPointFile.cpp
@@ -206,7 +206,7 @@ WayPointFile::Parse(Waypoints &way_points,
     if (reader.error())
       return false;
 
-    double filesize = std::max(reader.size(), 1l);
+    unsigned filesize = std::max(reader.size(), 1l);
 
     // Read through the lines of the file
     TCHAR *line;
@@ -224,7 +224,7 @@ WayPointFile::Parse(Waypoints &way_points,
     if (reader.error())
       return false;
 
-    double filesize = std::max(reader.size(), 1l);
+    unsigned filesize = std::max(reader.size(), 1l);
 
     // Read through the lines of the file
     TCHAR *line;
